discord platform to ts
- also some related file changes (Emote in globals etc)
