+++ classes/command.ts
@@ -1,18 +1,22 @@
 import { SupiDate, SupiError, isGenericRequestError, isGotRequestError, Counter, } from "supi-core"
 import type { Query, MetricConfiguration, MetricType } from "supi-core";
+import type { BaseMessageOptions } from "discord.js";
+
+type DiscordEmbeds = BaseMessageOptions["embeds"];
 
 import { TemplateWithoutId, TemplateDefinition } from "./template.js";
 
 import Banphrase from "./banphrase.js";
 import { Filter, ExecuteResult as FilterExecuteResult } from "./filter.js";
 import User from "./user.js";
-import { Channel, privateMessageChannelSymbol, type Emote } from "./channel.js";
+import { Channel, privateMessageChannelSymbol } from "./channel.js";
 import { Platform, type GetEmoteOptions } from "../platforms/template.js";
 import CooldownManager from "../utils/cooldown-manager.js";
 import { Language, LanguageParser } from "../utils/languages.js";
 
 import { whitespaceRegex } from "../utils/regexes.js";
 import config from "../config.json" with { type: "json" };
+import { Emote } from "../@types/globals.js";
 
 const COMMAND_PREFIX = config.modules.commands.prefix;
 const LINEAR_REGEX_FLAG = "--enable-experimental-regexp-engine";
@@ -63,6 +67,10 @@ export type StrictResult = {
 };
 type Result = StrictResult & {
 	reason?: string;
+	discord?: {
+		embeds?: DiscordEmbeds;
+		reactions: string[] | { emoji: string; }[];
+	};
 };
 
 export type Invocation = string;
